move yaml_config_writer.hpp to public includes
