+++ rviz_common/test/display_test.cpp
@@ -38,7 +38,7 @@
 #include "rviz_common/properties/color_property.hpp"
 #include "rviz_common/config.hpp"
 #include "rviz_common/yaml_config_reader.hpp"
-#include "../src/rviz_common/yaml_config_writer.hpp"
+#include "rviz_common/yaml_config_writer.hpp"
 
 #include "mock_display.hpp"
 #include "mock_display_group.hpp"
